demos/cadence/sim: fix timer task priority and CLOCK_HZ
Timer task with priority 1 results in timer inaccuracies in the presence
of tasks with higher priority. Timer resolution of 20ms is not
sufficient for audio demos.
Increase timer task priority to configMAX_PRIORITIES - 1 as in most other
ports.
Increase CLOCK_HZ to 1000 so that timer resolution become 1ms.
Increase CPU_CLOCK_HZ to 50MHz so that it is fast enough to do any
useful work with 1ms period timer.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>