xtensa port: fix timer setting in vPortSuppressTicksAndSleep
Timer setting in vPortSuppressTicksAndSleep can take a long time to
calculate new cycle number, and as a result it may set timer to a cycle
that is already in the past. When this happens FreeRTOS gets stuck in
the idle loop.

Do cycle delta calculation separately and add this delta to the current
timestamp in the vPortSuppressTicksAndSleep to fix this issue.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>