demos/cadence/sim: clean up make dependencies
Don't put 'mkdir' dependency at top level, this is not compatible with
parallel build and is not guaranteed to work. Instead add this
dependency to the targets that actually need build directory.

Replace implicit pattern rules with pattern rules. As a result
intermediate object files are not removed at the end of the build and
running build second time is a no-op.

No need to guess dependencies of the library, just build it
unconditionally, its makefile will take care of re-building stale parts.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>