xtensa port: generate offsets instead of hard-coding them
There's a number of places where assembly code need to access a field of
C structure. So far hardcoded/calculated offsets were used, but they are
hard to get right in the presence of multiple configuration options.

Instead make a list of valuable assembly constant and use it to compile
a program that generates a header with macro definitions for these
constants.

If there were no requirement of making it work on windows with only
xtensa tools it could be done with only target compiler and a sed script.
Without sed the solution is target compiler, linker and emulator.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>