xtensa port: count ticks separately
If xTaskIncrementTick is called while the scheduler is suspended then OS
tick counter update is delayed. vPortSuppressTicksAndSleep is always
called with suspended scheduler, so timer deadline in ccompare SR may
not correspong to the current OS tick. Count ticks separately in the
port and use that counter to calculate future timer deadlines.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>