xtensa port: fix timer IRQ clearing in NX port
Clear timer IRQ before the next timer deadline is set in the tickless
idle section that handles awakening by non-tick interrupt.
Since the timer interval set there may be arbitrary small take extra
care to clear any timer IRQ that could be pending before the ccount is
updated and not clear IRQ if it was set after the update.

This fixes timer_test2 on NX cores.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>